requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.30.6
repo-to-text==0.5.4
chromadb==0.6.3
sentence-transformers==3.4.1
pre-commit==4.2.0
pytest==8.3.5
gitpython==3.1.44
tree-sitter==0.24.0
tree-sitter-languages==1.10.2
pygments==2.19.1
tree-sitter-python==0.23.6
transformers==4.50.3
langchain
langchain-ollama
langchain-chroma
pandas
